implemented receive functionality
The received function is based on the interrupt vector USART_RX_vect which sotores the value of each received char into rx_buffer. This buffer was implemented in the UART lib without using the uint8-queue library because the uint8-queue library only allows for queues of only 10 positions. Also, it made sense for me to have the buffer for the received, implemented in the UART lib.