Re-wrote the state machine to be organized by events and then by state.
Since now state is a member of the timer struct and not of the state machine struct, the functionality of each event is repeated in several states, or it applies to all timers regardless if it is the current timer or not. This last one is the case of the SECOND_TICK event.