Move file specific to the simulator
This is done since PlatformIO insists on overwriting any launch.json
file. So we get unstaged changes each time pio is invoked...

This has the unfortunate downside that we can't use the launch file
unless we open the workspace with that directory as root.

See https://github.com/platformio/platformio-core/issues/3824