Go back to sleep if woken by an interrupt without user activity
Turns out this is actually a completely valid case!

The rotary encoder takes 4 substeps for each rotation, which thus makes
up a very small angle of rotation. And that makes it likely to happen
when just picking up the device, or shaking it lightly.

If this happens, we'll quite quickly go to sleep again when the
interrupts stop happening. That way we make sure to go down to low power
usage again.